Remove an advanced layer diff function
Noticed this while I was poking around in the runtime doing DB
work. The signature of this function makes me a bit uncomfortable
(why should we let people apply arbitrary diffs to layers? Seems
like a good way to break things...) and it's completely unused,
so let's just remove it.

[NO TESTS NEEDED] since this is a pure removal.

Signed-off-by: Matthew Heon <mheon@redhat.com>